Revert "Convert back to double"
This reverts commit a55967bdd6f3c401b8fdc5bb2855c9c37700c30a.

+++ include/SMVector.h
@@ -9,9 +9,9 @@
 // Vector struct
 namespace smtech1 {
     typedef struct SMVector {
-        double x;
-        double y;
-        double z;
+        int x;
+        int y;
+        int z;
     } SMVector;
 }
